Remove CLI ESLint suppressions (Fixes #2087)
