+++ packages/cli/src/ui/components/shared/text-buffer.ts
@@ -4,8 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-/* eslint-disable complexity, eslint-comments/disable-enable-pair -- Phase 5: legacy UI boundary retained while larger decomposition continues. */
-
 // Type re-exports — public API surface only
 export type {
   Direction,
