+++ packages/cli/src/ui/hooks/keyToAnsi.ts
@@ -4,11 +4,37 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-/* eslint-disable complexity, eslint-comments/disable-enable-pair -- Phase 5: legacy UI boundary retained while larger decomposition continues. */
-
 import type { Key } from '../contexts/KeypressContext.js';
 
 export type { Key };
+
+const KEY_TO_ANSI = new Map<string, string>([
+  ['up', '\x1b[A'],
+  ['down', '\x1b[B'],
+  ['right', '\x1b[C'],
+  ['left', '\x1b[D'],
+  ['escape', '\x1b'],
+  ['tab', '\t'],
+  ['backspace', '\x7f'],
+  ['delete', '\x1b[3~'],
+  ['home', '\x1b[H'],
+  ['end', '\x1b[F'],
+  ['pageup', '\x1b[5~'],
+  ['pagedown', '\x1b[6~'],
+  ['return', '\r'],
+]);
+
+function ctrlLetterToAnsi(name: string): string | null {
+  if (name >= 'a' && name <= 'z') {
+    return String.fromCharCode(name.charCodeAt(0) - 'a'.charCodeAt(0) + 1);
+  }
+  switch (name) {
+    case 'c':
+      return '\x03';
+    default:
+      return null;
+  }
+}
 
 /**
  * Translates a Key object into its corresponding ANSI escape sequence.
@@ -19,59 +45,15 @@ export type { Key };
  */
 export function keyToAnsi(key: Key): string | null {
   if (key.ctrl) {
-    // Ctrl + letter
-    if (key.name >= 'a' && key.name <= 'z') {
-      return String.fromCharCode(
-        key.name.charCodeAt(0) - 'a'.charCodeAt(0) + 1,
-      );
-    }
-    // Other Ctrl combinations might need specific handling
-    switch (key.name) {
-      case 'c':
-        return '\x03'; // ETX (End of Text), commonly used for interrupt
-      // Add other special ctrl cases if needed
-      default:
-        break;
-    }
+    return ctrlLetterToAnsi(key.name);
   }
 
-  // Arrow keys and other special keys
-  switch (key.name) {
-    case 'up':
-      return '\x1b[A';
-    case 'down':
-      return '\x1b[B';
-    case 'right':
-      return '\x1b[C';
-    case 'left':
-      return '\x1b[D';
-    case 'escape':
-      return '\x1b';
-    case 'tab':
-      return '\t';
-    case 'backspace':
-      return '\x7f';
-    case 'delete':
-      return '\x1b[3~';
-    case 'home':
-      return '\x1b[H';
-    case 'end':
-      return '\x1b[F';
-    case 'pageup':
-      return '\x1b[5~';
-    case 'pagedown':
-      return '\x1b[6~';
-    default:
-      break;
+  const mapped = KEY_TO_ANSI.get(key.name);
+  if (mapped !== undefined) {
+    return mapped;
   }
 
-  // Enter/Return
-  if (key.name === 'return') {
-    return '\r';
-  }
-
-  // If it's a simple character, return it.
-  if (!key.ctrl && !key.meta && key.sequence) {
+  if (!key.meta && key.sequence) {
     return key.sequence;
   }
