+++ packages/cli/src/ui/state/extensions.ts
@@ -4,8 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-/* eslint-disable complexity, eslint-comments/disable-enable-pair -- Phase 5: legacy UI boundary retained while larger decomposition continues. */
-
 import type { ExtensionUpdateInfo } from '../../config/extension.js';
 import { checkExhaustive } from '../../utils/checks.js';
 
@@ -108,21 +106,7 @@ export function extensionUpdatesReducer(
         batchChecksInProgress: state.batchChecksInProgress - 1,
       };
     case 'SCHEDULE_UPDATE':
-      return {
-        ...state,
-        // If there is a pre-existing scheduled update, we merge them.
-        scheduledUpdate: {
-          all: state.scheduledUpdate?.all === true || action.payload.all,
-          names: [
-            ...(state.scheduledUpdate?.names ?? []),
-            ...(action.payload.names ?? []),
-          ],
-          onCompleteCallbacks: [
-            ...(state.scheduledUpdate?.onCompleteCallbacks ?? []),
-            action.payload.onComplete,
-          ],
-        },
-      };
+      return { ...state, scheduledUpdate: mergeScheduledUpdate(state, action) };
     case 'CLEAR_SCHEDULED_UPDATE':
       return {
         ...state,
@@ -145,4 +129,21 @@ export function extensionUpdatesReducer(
     default:
       checkExhaustive(action);
   }
+}
+
+function mergeScheduledUpdate(
+  state: ExtensionUpdatesState,
+  action: Extract<ExtensionUpdateAction, { type: 'SCHEDULE_UPDATE' }>,
+): ScheduledUpdate {
+  return {
+    all: state.scheduledUpdate?.all === true || action.payload.all,
+    names: [
+      ...(state.scheduledUpdate?.names ?? []),
+      ...(action.payload.names ?? []),
+    ],
+    onCompleteCallbacks: [
+      ...(state.scheduledUpdate?.onCompleteCallbacks ?? []),
+      action.payload.onComplete,
+    ],
+  };
 }
