+++ packages/cli/src/utils/privacy/ConversationDataRedactor.ts
@@ -10,8 +10,6 @@
  * @pseudocode consumer-migration.md lines 10-15
  */
 
-/* eslint-disable complexity, eslint-comments/disable-enable-pair -- Phase 5: legacy CLI boundary retained while larger decomposition continues. */
-
 import type { IContent, ContentBlock } from '@vybestack/llxprt-code-core';
 import type { ITool } from '@vybestack/llxprt-code-providers';
 
@@ -168,19 +166,16 @@ export class ConversationDataRedactor {
 
     // SSH keys and certificates
     redacted = redacted.replace(
-      // eslint-disable-next-line sonarjs/regular-expr, sonarjs/slow-regex -- Static regex reviewed for lint hardening; bounded inputs preserve behavior.
       /\/[^"\s]*\.ssh\/[^"\s]*/g,
       '[REDACTED-SSH-PATH]',
     );
     redacted = redacted.replace(
-      // eslint-disable-next-line sonarjs/regular-expr, sonarjs/slow-regex -- Static regex reviewed for lint hardening; bounded inputs preserve behavior.
       /\/[^"\s]*\/id_rsa[^"\s]*/g,
       '[REDACTED-SSH-KEY-PATH]',
     );
 
     // Environment files
     redacted = redacted.replace(
-      // eslint-disable-next-line sonarjs/regular-expr, sonarjs/slow-regex -- Static regex reviewed for lint hardening; bounded inputs preserve behavior.
       /\/[^"\s]*\.env[^"\s]*/g,
       '[REDACTED-ENV-FILE]',
     );
@@ -204,23 +199,19 @@ export class ConversationDataRedactor {
 
     // Email addresses
     redacted = redacted.replace(
-      // eslint-disable-next-line sonarjs/regular-expr, sonarjs/slow-regex -- Static regex reviewed for lint hardening; bounded inputs preserve behavior.
       /[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}/g,
       '[REDACTED-EMAIL]',
     );
 
     // Phone numbers (basic patterns)
-    // eslint-disable-next-line sonarjs/regular-expr -- Static regex reviewed for lint hardening; behavior preserved.
     redacted = redacted.replace(/\b\d{3}-\d{3}-\d{4}\b/g, '[REDACTED-PHONE]');
     redacted = redacted.replace(
-      // eslint-disable-next-line sonarjs/regular-expr -- Static regex reviewed for lint hardening; behavior preserved.
       /\b\(\d{3}\)\s?\d{3}-\d{4}\b/g,
       '[REDACTED-PHONE]',
     );
 
     // Credit card numbers (basic pattern)
     redacted = redacted.replace(
-      // eslint-disable-next-line sonarjs/regular-expr -- Static regex reviewed for lint hardening; behavior preserved.
       /\b\d{4}[-\s]?\d{4}[-\s]?\d{4}[-\s]?\d{4}\b/g,
       '[REDACTED-CC-NUMBER]',
     );
@@ -331,35 +322,27 @@ export class ConversationDataRedactor {
     // Redact sensitive file paths
     const sensitivePatterns = [
       {
-        // eslint-disable-next-line sonarjs/regular-expr -- Static regex reviewed for lint hardening; behavior preserved.
         pattern: /\/home\/[^/]+\/\.ssh\/[^/]+/g,
         replacement: '[REDACTED-SSH-KEY-PATH]',
       },
       {
-        // eslint-disable-next-line sonarjs/regular-expr -- Static regex reviewed for lint hardening; behavior preserved.
         pattern: /\/home\/[^/]+\/\.aws\/[^/]+/g,
         replacement: '[REDACTED-AWS-CONFIG-PATH]',
       },
       {
-        // eslint-disable-next-line sonarjs/regular-expr -- Static regex reviewed for lint hardening; behavior preserved.
         pattern: /\/home\/[^/]+\/\.docker\/[^/]+/g,
         replacement: '[REDACTED-DOCKER-CONFIG-PATH]',
       },
       {
-        // eslint-disable-next-line sonarjs/regular-expr -- Static regex reviewed for lint hardening; behavior preserved.
         pattern: /\/Users\/[^/]+\/\.ssh\/[^/]+/g,
         replacement: '[REDACTED-SSH-KEY-PATH]',
       },
       {
-        // eslint-disable-next-line sonarjs/regular-expr -- Static regex reviewed for lint hardening; behavior preserved.
         pattern: /\/Users\/[^/]+\/\.aws\/[^/]+/g,
         replacement: '[REDACTED-AWS-CONFIG-PATH]',
       },
-      // eslint-disable-next-line sonarjs/regular-expr, sonarjs/slow-regex -- Static regex reviewed for lint hardening; bounded inputs preserve behavior.
       { pattern: /.*\.env.*$/g, replacement: '[REDACTED-SENSITIVE-PATH]' },
-      // eslint-disable-next-line sonarjs/regular-expr, sonarjs/slow-regex -- Static regex reviewed for lint hardening; bounded inputs preserve behavior.
       { pattern: /.*secret.*$/gi, replacement: '[REDACTED-SENSITIVE-PATH]' },
-      // eslint-disable-next-line sonarjs/regular-expr, sonarjs/slow-regex -- Static regex reviewed for lint hardening; bounded inputs preserve behavior.
       { pattern: /.*key.*$/gi, replacement: '[REDACTED-SENSITIVE-PATH]' },
     ];
 
@@ -377,19 +360,16 @@ export class ConversationDataRedactor {
 
     // Redact export statements with sensitive values
     redacted = redacted.replace(
-      // eslint-disable-next-line sonarjs/regular-expr -- Static regex reviewed for lint hardening; behavior preserved.
       /export\s+[A-Z_]+\s*=\s*['"]\s*sk-[a-zA-Z0-9]+\s*['"]/g,
       'export [REDACTED-API-KEY]',
     );
     redacted = redacted.replace(
-      // eslint-disable-next-line sonarjs/regular-expr -- Static regex reviewed for lint hardening; behavior preserved.
       /export\s+[A-Z_]+\s*=\s*['"]\s*[a-zA-Z0-9+/]+=*\s*['"]/g,
       'export [REDACTED-TOKEN]',
     );
 
     // Redact curl commands with authorization headers
     redacted = redacted.replace(
-      // eslint-disable-next-line sonarjs/regular-expr, sonarjs/slow-regex -- Static regex reviewed for lint hardening; bounded inputs preserve behavior.
       /curl\s+.*-H\s+['"]Authorization:\s*Bearer\s+[^'"]+['"]/g,
       'curl [REDACTED-AUTH-HEADER]',
     );
@@ -440,21 +420,18 @@ export class ConversationDataRedactor {
       },
       {
         name: 'email_addresses',
-        // eslint-disable-next-line sonarjs/regular-expr -- Static regex reviewed for lint hardening; behavior preserved.
         pattern: /\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Z|a-z]{2,}\b/g,
         replacement: '[REDACTED-EMAIL]',
         enabled: this.redactionConfig.redactEmails,
       },
       {
         name: 'passwords',
-        // eslint-disable-next-line sonarjs/regular-expr -- Static regex reviewed for lint hardening; behavior preserved.
         pattern: /(?:password|pwd|pass)[=:\s]+[^\s\n\r]+/gi,
         replacement: 'password=[REDACTED]',
         enabled: this.redactionConfig.redactCredentials,
       },
       {
         name: 'generic_api_keys',
-        // eslint-disable-next-line sonarjs/regular-expr -- Static regex reviewed for lint hardening; behavior preserved.
         pattern: /api[_-]?key["\s]*[:=]["\s]*[a-zA-Z0-9-_]{16,}/gi,
         replacement: 'api_key: "[REDACTED-API-KEY]"',
         enabled: this.redactionConfig.redactApiKeys,
